feat: register and delete categories
