+++ handler/cadastro_categorias_servicos.go
@@ -5,6 +5,7 @@ import (
 
 	"github.com/gin-gonic/gin"
 	"github.com/lcardelli/fornecedores/schemas"
+	"gorm.io/gorm"
 )
 
 func CadastroCategoriaHandler(c *gin.Context) {
@@ -23,6 +24,82 @@ func CadastroCategoriaHandler(c *gin.Context) {
 		"user":       user,
 		"activeMenu": "cadastro-categoria",
 	})
+}
+
+func CreateCategoryHandler(c *gin.Context) {
+	var category schemas.SupplierCategory
+	if err := c.ShouldBindJSON(&category); err != nil {
+		c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
+		return
+	}
+
+	if err := db.Create(&category).Error; err != nil {
+		c.JSON(http.StatusInternalServerError, gin.H{"error": "Erro ao criar categoria"})
+		return
+	}
+
+	c.JSON(http.StatusCreated, category)
+}
+
+func UpdateCategoryHandler(c *gin.Context) {
+	id := c.Param("id")
+	var category schemas.SupplierCategory
+
+	if err := db.First(&category, id).Error; err != nil {
+		c.JSON(http.StatusNotFound, gin.H{"error": "Categoria não encontrada"})
+		return
+	}
+
+	if err := c.ShouldBindJSON(&category); err != nil {
+		c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
+		return
+	}
+
+	if err := db.Save(&category).Error; err != nil {
+		c.JSON(http.StatusInternalServerError, gin.H{"error": "Erro ao atualizar categoria"})
+		return
+	}
+
+	c.JSON(http.StatusOK, category)
+}
+
+func DeleteCategoryHandler(c *gin.Context) {
+	id := c.Param("id")
+	
+	// Verifica se o ID é válido
+	if id == "" {
+		c.JSON(http.StatusBadRequest, gin.H{"error": "ID da categoria não fornecido"})
+		return
+	}
+
+	var category schemas.SupplierCategory
+	result := db.First(&category, id)
+
+	if result.Error != nil {
+		if result.Error == gorm.ErrRecordNotFound {
+			c.JSON(http.StatusNotFound, gin.H{"error": "Categoria não encontrada"})
+		} else {
+			c.JSON(http.StatusInternalServerError, gin.H{"error": "Erro ao buscar categoria"})
+		}
+		return
+	}
+
+	if err := db.Delete(&category).Error; err != nil {
+		c.JSON(http.StatusInternalServerError, gin.H{"error": "Erro ao deletar categoria"})
+		return
+	}
+
+	c.JSON(http.StatusOK, gin.H{"message": "Categoria deletada com sucesso"})
+}
+
+func ListCategoriesHandler(c *gin.Context) {
+	var categories []schemas.SupplierCategory
+	if err := db.Find(&categories).Error; err != nil {
+		c.JSON(http.StatusInternalServerError, gin.H{"error": "Erro ao buscar categorias"})
+		return
+	}
+
+	c.JSON(http.StatusOK, categories)
 }
 
 func CadastroServicoHandler(c *gin.Context) {
