--- handler/category_service.go
@@ -1,48 +1,0 @@
-package handler
-
-import (
-	"net/http"
-
-	"github.com/gin-gonic/gin"
-	"github.com/lcardelli/fornecedores/schemas"
-)
-
-func CreateCategoryHandler(c *gin.Context) {
-	var category schemas.SupplierCategory
-	if err := c.ShouldBindJSON(&category); err != nil {
-		c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
-		return
-	}
-
-	if err := db.Create(&category).Error; err != nil {
-		c.JSON(http.StatusInternalServerError, gin.H{"error": "Erro ao criar categoria"})
-		return
-	}
-
-	c.JSON(http.StatusCreated, category)
-}
-
-func CreateServiceHandler(c *gin.Context) {
-	var service schemas.Service
-	if err := c.ShouldBindJSON(&service); err != nil {
-		c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
-		return
-	}
-
-	if err := db.Create(&service).Error; err != nil {
-		c.JSON(http.StatusInternalServerError, gin.H{"error": "Erro ao criar serviço"})
-		return
-	}
-
-	c.JSON(http.StatusCreated, service)
-}
-
-func ListCategoriesHandler(c *gin.Context) {
-	var categories []schemas.SupplierCategory
-	if err := db.Find(&categories).Error; err != nil {
-		c.JSON(http.StatusInternalServerError, gin.H{"error": "Erro ao buscar categorias"})
-		return
-	}
-
-	c.JSON(http.StatusOK, categories)
-}
